Refactor field resolver for Learner, Activity, and Event.
