+++ server/src/resolver/index.ts
@@ -4,14 +4,19 @@ import { EventResolver } from "./event/EventResolver.js";
 import { LearnerResolver } from "./learner/LearnerResolver.js";
 import { LearningFileResolver } from "./learningFile/LearningFileResolver.js";
 import { TechniqueResolver } from "./technique/TechniqueResolver.js";
-
+import { LearnerFieldResolver } from "./learner/LearnerFieldResolver.js"
+import { ActivityFieldResolver } from "./activity/ActivityFieldResolver.js";
+import { EventFieldResolver } from "./event/EventFieldResolver.js";
 
 const resolvers: NonEmptyArray<Function> = [
   LearnerResolver,
   ActivityResolver,
   TechniqueResolver,
   LearningFileResolver,
-  EventResolver
+  EventResolver,
+  LearnerFieldResolver,
+  EventFieldResolver,
+  ActivityFieldResolver,
 ]
 
 export default resolvers;
